feat(scl): add benchmark
Benchmark run on my machine for the `git` repository:

```
assign_ids              time:   [2.0980 s 2.1244 s 2.1533 s]
                        change: [-0.8091% +1.2793% +3.3889%] (p = 0.27 > 0.05)
                        No change in performance detected.
```

There are ~66k commits in my copy of the repository, so this is ~31k commits per second. Note that the filesystem caches are warmed up as part of this benchmarking; we can expect slower real-world speeds.